* 'main' of https://github.com/engineering-thesis-PJATK/EngineeringThesis-backend-dev: (35 commits)
  SCAFFOLD - FINAL
  Employee integration tests
  Repair commit
  Added encryption of changing password
  Added method to send email with generated random password and method to changing password in database - without encryption
  Added tests for kanban repository
  Added controller and method for updating kanban elements status
  Synchronization between frontend and backend to kanban element
  Added repository for kanban elements. Added method for tasks and for tickets to send elements for kanaban. Added controller for kanban
  Added dto class for kanban and method to return list of tickets for kanban elements
  Changed structure of project
  Added UpdateCustomer_PassValid_ShouldUpdateCustomerInDatabase method for customer repository integration tests
  Added AddNewCustomer_WithNotValidCompanyId_ShouldThrowArgumentError method for customer repository tests and change add method in customer repository - added company existing validation
  Added attributes to method AddNewCustomer_WithNotValidEmail_ShouldThrowValidationError
  Added test method for customer repository test add new customer with no valid email
  Added test method for customer repository add new customer with pass valid
  Added class with integration tests for customers
  Added checking of existing company and customer in controller to return specific response
  Added update method for Customer
  Added Customer handler with method to checking unique of customer email. Added customer validation class
  ...

# Conflicts:
#	OneBan TMS/.idea/.idea.OneBan TMS/.idea/workspace.xml
#	OneBan TMS/obj/Debug/net5.0/OneBan TMS.csproj.AssemblyReference.cache
#	OneBan TMS/obj/rider.project.restore.info

+++ OneBan TMS/Repository/AddressRepository.cs
@@ -4,8 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using OneBan_TMS.Interfaces;
+using OneBan_TMS.Interfaces.Repositories;
 using OneBan_TMS.Models;
 using OneBan_TMS.Models.DTOs;
+using OneBan_TMS.Models.DTOs.Address;
 
 namespace OneBan_TMS.Repository
 {
@@ -49,6 +51,8 @@ namespace OneBan_TMS.Repository
                 .Addresses
                 .Where(x => x.AdrId == addressId)
                 .SingleOrDefaultAsync();
+            if (address is null)
+                throw new ArgumentException("Address not exists");
             address.AdrTown = updatedAddress.AdrTown;
             address.AdrStreet = updatedAddress.AdrStreet;
             address.AdrStreetNumber = updatedAddress.AdrStreetNumber;
@@ -62,8 +66,10 @@ namespace OneBan_TMS.Repository
                     .Addresses
                     .Where(x => x.AdrId == addressId)
                     .SingleOrDefaultAsync();
-                _context.Addresses.Remove(address);
-                await _context.SaveChangesAsync();
+            if (address is null)
+                throw new ArgumentException("Address not exists");
+            _context.Addresses.Remove(address);
+            await _context.SaveChangesAsync();
         }
 
         public async Task<bool> ExistsAddress(int addressId)
